feat: solve day7

+++ src/AdventOfCode/Common/Day.cs
@@ -2,6 +2,6 @@ namespace AdventOfCode.Common;
 
 public abstract class Day
 {
-    public abstract int SolvePuzzle1(string[] input);
-    public abstract int SolvePuzzle2(string[] input);
+    public abstract long SolvePuzzle1(string[] input);
+    public abstract long SolvePuzzle2(string[] input);
 }
